platform/mellanox: mlxreg-hotplug: Add line card event callbacks support for modular system
Add event callback field to 'mlxreg_hotplug_device' structure. These
callback functions are set by a probing routine of a driver, which is
supposed to perform some action when particular event is received. The
probing routine is also supposed to set the handle of the object
associated with the events. For example power on/ready/shutdown events
for the line card in slot #n.
If callback function is set, it will be called by ‘mlxreg-hotplug’
driver, when it receives signals related to line card state change.

The next signals could be received  for line card in particular slot:
- Line card was inserted or removed.
- Line card passed security signature validation by hardware.
- Line card has been powered on / off.
- Line card passed hardware-firmware synchronization handshake.
- Line card received ready event, indicating its PHYs ready / unready
  states.
- Line card received enable event indicating that line card’s FPGA and
  PHYs have been enabled.
- Line card received ready event indicating that ASIC firmware is ready
  for handling network, hardware monitoring and thermal interfaces
  associated with PHYs and ports found on this line card.
- Line card thermal shutdown event indicating that line card has some
  thermal issue and should be powered off.

Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>