mlxsw: minimal: Extend internal structures to support port areas
Introduce intermediate level for ports area. Currently all ports are
located at the main board. Port related interfaces are created during
driver initialization and removed during driver de-initialization.

For line cards in modular system the ports are to be associated with the
specific line card. Interfaces for the ports found on line card should
be created whenever new line card is available (inserted, validated,
powered, enabled and activated) and removed, when line card is getting
unavailable.

The ports found on the line card #n are accessed by setting slot index
to #n, while for access to the ports found on the main board slot index
should be set to default value zero. Each ports area contains the set
of network interfaces associated with particular slot index.

Thus introduction of ports area allows to use the same APIs for the main
board and line cards, by setting slot index argument.

Extend mlxsw_m_port_create() with 'slot_index' argument. The purpose is
to associate line card slot location with the ports found on this line
card.

For main board it'll always be zero, for line cards it'll be set to the
physical slot number at which line card is located. For now the slot
index is always 0.
That is, ports are still not instantiated from line cards.

Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>