platform/mellanox: mlxreg-io: Fix argument base in kstrtou32() call
Change kstrtou32() argument 'base' to zero instead of 'len'.
It works by chance for setting one bit value, but it is not supposed to
work in case value passed to mlxreg_io_attr_store() is greater than 1.

It works for example, for:
echo 1 > /sys/devices/platform/mlxplat/mlxreg-io/hwmon/.../jtag_enable
But it will fail for:
echo n > /sys/devices/platform/mlxplat/mlxreg-io/hwmon/.../jtag_enable,
where n > 1.

The flow for input buffer conversion is as below:
_kstrtoull(const char *s, unsigned int base, unsigned long long *res)
calls:
rv = _parse_integer(s, base, &_res);

For the second case in _parse_integer() 'val' is converted to 'n' and
'base' equal or greater then '2' (for 2 it will be be 0x32 0x0a, for 3
0x33 0x0a, etcetera). And _parse_integer() will exit with result zero
(rv):
	rv = 0;
	while (1) {
		...
		if (val >= base)
			break;
		...
		rv++;
		...
	}

And _kstrtoull() in their turn will fail:
	if (rv == 0)
		return -EINVAL;

Fixes: 5ec4a8ace06c: (platform/mellanox: Introduce support for Mellanox register access driver)
Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>